Got the hours display working
With some tricks to avoid huge duplicates of writing each units of the time onto the string, I managed to do it right.

The only problem is that it consumes a HUGE 11 tiles, for the timer itself when using hours. The test sprite would therefore draw a total 12 tiles, which is bigger than the highest maximum number of tiles allowed in the sprite memory settings (which is 10 for settings $00 and $0A). I might be wrong, let me install the NMSTL patch...