update
TestSprite_NumberDisplay: now works without the need of the share subroutines.

SpriteHUDTest: Fix a bug where uninstalling causes the game to crash, Caused by the incsrc that included the code is being outside the if statement, so if you uninstall, these included code won't be "auto-cleaned" out, causing leftover code somewhere in the game's code.